==> code/Step_1a-Import_and_clean_data.ipynb <==
# Notebook Intro:

In this notebook, I import the training data and perform basic clean (review nulls and datatypes). 

I then output the cleaned data to a csv so other notebooks can use the clean data to create a model.

Note that the data dictionary is [here](http://jse.amstat.org/v19n3/decock/DataDocumentation.txt).

## import packages and data

# imports
import numpy as np
import pandas as pd

# import training data
filepath = '../datasets/Original/train.csv'

train = pd.read_csv(filepath)
train.head()

## Check on general features of the data:

* review datatypes 
* review if any data is missing

train.info()

train['Mas Vnr Area'].value_counts()

train[train['Lot Frontage'].notna()]['Lot Frontage'].sort_values()
#based on results from this - the na could be a 0 sized lot rather than a missing value - it's not clear

**Analyze the following columns with data that is a float datatype and some are NA**
* if the data has a 0 this means the NA value is likely an error and wouldn't correspond to 0

train['BsmtFin SF 1'].value_counts()

train['BsmtFin SF 2'].value_counts()

train['Bsmt Unf SF'].value_counts()

train['Total Bsmt SF'].value_counts()

train['Bsmt Full Bath'].value_counts()

train['Bsmt Half Bath'].value_counts()

train['Garage Cars'].value_counts()

train['Garage Area'].value_counts()

The following column, 'Garage Yr Blt' does not have zeroes, so I review to see if the NA's correspond to when the Garage type is NaN (which means that the property has no garage).  

The following two lines of code and analysis show that 'Garage Yr Blt' being NA corresponds to no garage being built. So I'll convert the NA's in the 'Garage Yr Blt' column to another value.  Since this is a float column, I'll convert the Na's to 0's and make a note of it in the file.

train[train['Garage Type'].isnull()]['Garage Yr Blt'].sort_values()

train[train['Garage Type'].notna()]['Garage Yr Blt'].sort_values()

### Clean null values
A few columns have almost no data, so look into what NA means:

|Column Name|  total non-null entries| approximate % of data| what NA means| convert/remove NA
| :-: | :-: | :-: | :- |:- |
|'Pool QC'| 9 | 0.4%| per dictionary, NA means No Pool, so keep this column (NA has meaning)|convert to none
|'Misc Feature'| 65 | 3.2%| per dictionary, NA means none, so can keep this column (NA has meaning)|convert to none
|'Alley'| 140 | 6.8%| per dictionary, NA means no alley access NA has meaning, so keep)|convert to none
|'Fence'| 400 | 9.5% |per dictionary, NA means No Fence, so keep this column (NA has meaning)|convert to none

**all columns with high number of NA values have meaning to the NA, so keeping these columns.**



A few columns have some entries missing, so look into what NA means for these:

|index| Column Name| non-null entries | data type in column| analysis on what the NA means| convert/remove NA entries|
| :-: | :-: | :-: | :-: | :-: |:-: |
| 4  | Lot Frontage   |  1721 non-null |  float64  |   based on analysis above, na could be values of 0| because this is a data type float category, convert to 0 with note in file that 0 means it was previously missing data
| 26 | Mas Vnr Type   |  2029 non-null |  object   |   per dictionary - none is entered if no masonry, so Na is an error| remove
| 27 | Mas Vnr Area   |  2029 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 31 | Bsmt Qual      |  1996 non-null |  object   |   per dictionary - NA means no basement | convert to none
| 32 | Bsmt Cond      |  1996 non-null |  object   |   per dictionary - NA means no basement | convert to none
| 33 | Bsmt Exposure  |  1993 non-null |  object   |   per dictionary - NA means no basement | convert to none
| 34 | BsmtFin Type 1 |  1996 non-null |  object   |   per dictionary - NA means no basement | convert to none
| 35 | BsmtFin SF 1   |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 36 | BsmtFin Type 2 |  1995 non-null |  object   |   per dictionary - NA means no basement | convert to none
| 37 | BsmtFin SF 2   |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 38 | Bsmt Unf SF    |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 39 | Total Bsmt SF  |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 48 | Bsmt Full Bath |  2049 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 49 | Bsmt Half Bath |  2049 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 58 | Fireplace Qu   |  1051 non-null |  object   |   per dictionary - NA means no fireplace | convert to none
| 59 | Garage Type    |  1938 non-null |  object   |   per dictionary - NA means no garage   | convert to none
| 60 | Garage Yr Blt  |  1937 non-null |  float64  |   based on analysis - NA means no garage|because this is a data type float category, convert to 0 with note in file that 0 means it was previously missing data
| 61 | Garage Finish  |  1937 non-null |  object   |   per dictionary - NA means no garage   | convert to none
| 62 | Garage Cars    |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove 
| 63 | Garage Area    |  2050 non-null |  float64  |   per analysis - 0 is entered as an area, so NA is an error | remove
| 64 | Garage Qual    |  1937 non-null |  object   |   per dictionary - NA means no garage   | convert to none
| 65 | Garage Cond    |  1937 non-null |  object   |   per dictionary - NA means no garage   | convert to none
 


**Prior to removing null values that don't have meaning (which affects 30 to 40 rows total or < 2%), 
convert null values as appropriate based on review above**

# function to convert missing values to a selected value

def na_to_replacement_value(dataframe,list_of_column_names,replaced_value):
    for column in list_of_column_names:
        dataframe[column] = dataframe[column].replace(np.nan,replaced_value)          
    return dataframe

# perform function on the following columns and replace with none
# note that all columns have type object 

columns_na_to_none = ['Pool QC','Misc Feature','Alley','Fence','Bsmt Qual','Bsmt Cond','Bsmt Exposure','BsmtFin Type 1','BsmtFin Type 2','Fireplace Qu','Garage Type','Garage Finish','Garage Qual','Garage Cond']
train = na_to_replacement_value(train,columns_na_to_none,'none')

# convert the columns with missing values and float data types to 0

columns_na_to_0 = ['Lot Frontage','Garage Yr Blt']
train = na_to_replacement_value(train,columns_na_to_0,0)


train.info()

# after analysis remaining null values are likely error, so remove from dataframe

train_clean = train.dropna()

train_clean.info()

percent_data_lost = round((len(train)-len(train_clean))/len(train),4)*100
print(f'By eliminating the missing rows, the data frame went from {len(train)} to {len(train_clean)} rows, or it lost {percent_data_lost}% of the data')

This is a reasonable amount of data to lose. So this is fine.

# export clean data
filepath = '../datasets/interim_files/train_clean.csv'

train_clean.to_csv(filepath,index=False)


# Notebook Summary of Work:

In this notebook, I imported the training data and performed basic clean up (review nulls and datatypes). 

I then output the cleaned data to a csv so other notebooks can use the clean data to create a model.
