==> code/Step_1b-import_and_clean_test_data_nodropping_NA_in_rows.ipynb <==
# Notebook Intro:

In this notebook, I import the test data and perform basic clean (review nulls and datatypes), similar to what was performed in notebook 1a. 

I then output the cleaned data to a csv so other notebooks can use the clean test data for models for final submission.

Note that the data dictionary is [here](http://jse.amstat.org/v19n3/decock/DataDocumentation.txt).

# imports

import numpy as np
import pandas as pd

# import test data
filepath = '../datasets/Original/test.csv'

test = pd.read_csv(filepath)
test.head()

test.info()

# function to convert missing values to a selected value

def na_to_replacement_value(dataframe,list_of_column_names,replaced_value):
    for column in list_of_column_names:
        dataframe[column] = dataframe[column].replace(np.nan,replaced_value)          
    return dataframe

# perform function on the following columns and replace with none
# note that all columns have type object 

columns_na_to_none = ['Pool QC','Misc Feature','Alley','Fence','Bsmt Qual','Bsmt Cond','Bsmt Exposure','BsmtFin Type 1','BsmtFin Type 2','Fireplace Qu','Garage Type','Garage Finish','Garage Qual','Garage Cond']
test = na_to_replacement_value(test,columns_na_to_none,'none')

# convert the columns with missing values and float data types to 0

columns_na_to_0 = ['Lot Frontage','Garage Yr Blt']
test = na_to_replacement_value(test,columns_na_to_0,0)


test.info()

# dropping for right now while need to get 878 entries

#dropping for right now while need to get 878 entries
#test_clean= test.dropna()

test_clean = test

test_clean.info()

#note that I will need to impute data in future if any columns contain null values for features that I'm analyzing.

# export clean data
filepath = '../datasets/interim_files/test_clean.csv'

test_clean.to_csv(filepath,index=False)

# Notebook Summary of Work:

In this notebook, I import the test data and perform basic clean (review nulls and datatypes), similar to what was performed in notebook 1a. 

I then output the cleaned data to a csv so other notebooks can use the clean test data for models for final submission.
