==> code/Step_2b-Create_Prediction_1_using_top_5_features_and_LinReg_no_scale.ipynb <==
# Notebook Intro:

In this notebook, I use my cleaned training data along with the five features I identified in notebook 2a as having the best r2 values wrt Sale Price and create a linear regression model.

I first do a train/test split on the training data, then I fit the linear regression model to the train split of the data, verify the model is reasonable by doing a cross-val score as well as running is on the test split of the data.

Finally, I use linear regression model that I fit on the train split of the training data on the test data and export this for **Prediction 1**.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split, cross_val_score

# import cleaned info
filepath = '../datasets/interim_files/train_clean.csv'

df = pd.read_csv(filepath)
df.head()

# Create matricies

features = ['Overall Qual','Gr Liv Area','Garage Area','Garage Cars','Total Bsmt SF']

X = df[features]
y = df['SalePrice']

# Train-Test-Split

X_train, X_test, y_train, y_test = train_test_split(X,y, random_state=42)

# Instantiate Model

lr = LinearRegression()

# cross val score

cross_val_score(lr,X_train,y_train).mean()

# fit model since ok cross val score

lr.fit(X_train,y_train)

# train score

lr.score(X_train,y_train)

# similar to cross val score, so keep model

lr.score(X_test, y_test)

# test and train scores - so apply to test data

## apply model to test data





# import cleaned test info
filepath = '../datasets/interim_files/test_clean.csv'

testdata = pd.read_csv(filepath)
testdata.head()

testdata.info()

# Create matricies for test data

features = ['Overall Qual','Gr Liv Area','Garage Area','Garage Cars','Total Bsmt SF']

X_testdata = testdata[features]

#verify no nulls in the dataset
X_testdata.info()

# Make Predictions

y_pred = lr.predict(X_testdata)

y_pred_df = pd.DataFrame(y_pred)
y_pred_df = y_pred_df.rename(columns={0:'SalePrice'})
y_pred_df.head()

final_preds = pd.concat([testdata['Id'],y_pred_df], axis=1 )
final_preds.info()

#export predictions
filepath = '../datasets/submissions/prediction1.csv'

final_preds.to_csv(filepath, index=False)

# Notebook Summary of Work:

In this notebook, I use my cleaned training data along with the five features I identified in notebook 2a as having the best r2 values wrt Sale Price and create a linear regression model.

I first do a train/test split on the training data, then I fit the linear regression model to the train split of the data, verify the model is reasonable by doing a cross-val score as well as running is on the test split of the data.

Finally, I use linear regression model that I fit on the train split of the training data on the test data and export this for **Prediction 1**.
