==> code/Step_3b-Create_Prediction_2__using_categorical_and_continuous_features.ipynb <==
# Notebook Intro:

In this notebook, I combine my selected dummy columns from Notebook 3a, and combine this with continuous features that have an r2 above .2 wrt Sale Price and create a linear regression model.

I first do a train/test split on the training data, then I fit the linear regression model to the train split of the data, verify the model is reasonable by doing a cross-val score as well as running is on the test split of the data.

Finally, I modify the training data similar to the testing data (creating the dummy columns and selecting only the cotinuous features that I used in the training data), and I use the linear regression model that I fit on the train split of the training data on the test data and export this for **Prediction 2**.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LinearRegression

# import cleaned training data
filepath = '../datasets/interim_files/train_clean.csv'

df = pd.read_csv(filepath)

# import dummy columns with r2 over.2
filepath = '../datasets/interim_files/dummiesbest.csv'

df_dummies = pd.read_csv(filepath)

# continuous features - per the data dictionary
continuous_features = ['Lot Frontage','Lot Area','Mas Vnr Area','BsmtFin SF 1','BsmtFin SF 2','Bsmt Unf SF','Total Bsmt SF','1st Flr SF','2nd Flr SF','Low Qual Fin SF','Gr Liv Area','Garage Area','Wood Deck SF','Open Porch SF','Enclosed Porch','3Ssn Porch','Screen Porch','Pool Area','Misc Val']

# nominal features - per the data dictionary
nominal_features = ['PID','MS SubClass','MS Zoning','Street','Alley','Land Contour','Lot Config','Neighborhood','Condition 1','Condition 2','Bldg Type','House Style','Roof Style','Roof Matl','Exterior 1st','Exterior 2nd','Mas Vnr Type','Foundation','Heating','Central Air','Garage Type','Misc Feature','Sale Type']

# discrete features 
discrete_features = ['Year Built','Year Remod/Add','Bsmt Full Bath','Bsmt Half Bath','Full Bath','Half Bath','Bedroom AbvGr','Kitchen AbvGr','TotRms AbvGrd','Fireplaces','Garage Yr Blt','Garage Cars','Mo Sold','Yr Sold']

# Ordinal Features
ordinal_features = ['Lot Shape','Utilities','Land Slope','Overall Qual','Overall Cond','Exter Qual','Exter Cond','Bsmt Qual','Bsmt Cond','Bsmt Exposure','BsmtFin Type 1','BsmtFin Type 2','Heating QC','Electrical','Kitchen Qual','Functional','Fireplace Qu','Garage Finish','Garage Qual','Garage Cond','Paved Drive','Pool QC','Fence']

# look at continuous variable

df[continuous_features]

continuous_features_with_sale = pd.concat([df[continuous_features],df['SalePrice']], axis = 1)
np.abs(continuous_features_with_sale.corr()['SalePrice'])

continuous_features_with_sale.corr()['SalePrice'][np.abs(continuous_features_with_sale.corr()['SalePrice']) >.2]

final_features_continuous = continuous_features_with_sale.corr()['SalePrice'][np.abs(continuous_features_with_sale.corr()['SalePrice']) >.2]
final_features_continuous.index

df_continuous = df[final_features_continuous.index].copy()
df_continuous.drop(columns = 'SalePrice', inplace=True)

# combine nominal and continuous features

X = pd.concat([df_dummies,df_continuous], axis=1)
y = df['SalePrice']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42)

lr = LinearRegression()

cross_val_score(lr, X_train, y_train).mean()

lr.fit(X_train, y_train)

lr.score(X_train, y_train), lr.score(X_test, y_test)

# apply model to test data

# import cleaned test info
filepath = '../datasets/interim_files/test_clean.csv'

testdata = pd.read_csv(filepath)
testdata.head()

testdata.shape

# columns used in dummies (all the nominal columns that had relevant dummy values) - from notebook 3a
combined_columns = ['MS SubClass', 'MS Zoning', 'Land Contour', 'Neighborhood',
       'House Style', 'Roof Style', 'Exterior 1st', 'Exterior 2nd',
       'Mas Vnr Type', 'Foundation', 'Central Air', 'Garage Type',
       'Sale Type']

# get dummy columns from categorical variables
df_long_testdata = pd.get_dummies(testdata, columns = combined_columns)

df_to_review_testdata = df_long_testdata[df_dummies.columns]

# get continuous columns
df_continuous_testdata = testdata[final_features_continuous.index[:-1]]

df_continuous_testdata.shape

df_to_review_testdata.shape

X_testdata = pd.concat([df_to_review_testdata,df_continuous_testdata], axis=1)

X_testdata.info()

X_testdata.fillna(0, inplace=True)

y_pred = lr.predict(X_testdata)

y_pred_df = pd.DataFrame(y_pred)
y_pred_df = y_pred_df.rename(columns={0:'SalePrice'})
y_pred_df.head()

final_preds = pd.concat([testdata['Id'],y_pred_df], axis=1 )
final_preds.info()

#export predictions
filepath = '../datasets/submissions/prediction2.csv'

final_preds.to_csv(filepath, index=False)

# Notebook Summary:

In this notebook, I combine my selected dummy columns from Notebook 3a, and combine this with continuous features that have an r2 above .2 wrt Sale Price and create a linear regression model.

I first do a train/test split on the training data, then I fit the linear regression model to the train split of the data, verify the model is reasonable by doing a cross-val score as well as running is on the test split of the data.

Finally, I modify the training data similar to the testing data (creating the dummy columns and selecting only the cotinuous features that I used in the training data), and I use the linear regression model that I fit on the train split of the training data on the test data and export this for **Prediction 2**.
