==> code/Step_6-Combined_df_with_all_features_use_lr_lasso_ridge.ipynb <==
# Notebook Intro:

In this notebook, I use my cleaned training data, then look at all possible features, using all continuous and discrete features, importing the updated ordinal features (the ones with numerical values) from notebook 4, and then creating dummy columns for all nominal features with dropping the first dummy column for each one.

I perform a train/test split on this data, fit a scaler on the train split and then transform both the training and test splits.

I then fit a lasso model on the train split and run the model on both the training and test split.  I get good scores back, so I then perform the scaling and fitting of the lasso model to the entire set of training data.  I then use this updated lasso model on the test data to create **prediction 3**.  Note that while fitting the lasso model to the test data, I realize that some dummy columns don't align with the training and test data, so I combine these, then separate them so that both the training and test data will have the same columns.  All values that are null (because they didn't previously exist, are replaced with 0).

I export the lasso model on the training data for use in other notebooks.


I then fit a ridge model on the train split and run the model on both the training and test split.  However, it has similar performance to the lasso model, and a lot more features, so I don't use it to create any predictions with the test data.

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, RidgeCV, Lasso, LassoCV
from sklearn.pipeline import Pipeline

# import cleaned training data
filepath = '../datasets/interim_files/train_clean.csv'
df = pd.read_csv(filepath)

#import ordinal features
filepath = '../datasets/interim_files/training_updated_ordinal_features.csv'
df_ord = pd.read_csv(filepath)

# From Step_3- Notebook

# continuous features - per the data dictionary
continuous_features = ['Lot Frontage','Lot Area','Mas Vnr Area','BsmtFin SF 1','BsmtFin SF 2','Bsmt Unf SF','Total Bsmt SF','1st Flr SF','2nd Flr SF','Low Qual Fin SF','Gr Liv Area','Garage Area','Wood Deck SF','Open Porch SF','Enclosed Porch','3Ssn Porch','Screen Porch','Pool Area','Misc Val']

# nominal features - remove PID
nominal_features = ['MS SubClass','MS Zoning','Street','Alley','Land Contour','Lot Config','Neighborhood','Condition 1','Condition 2','Bldg Type','House Style','Roof Style','Roof Matl','Exterior 1st','Exterior 2nd','Mas Vnr Type','Foundation','Heating','Central Air','Garage Type','Misc Feature','Sale Type']

#'PID',

# discrete features 
discrete_features = ['Year Built','Year Remod/Add','Bsmt Full Bath','Bsmt Half Bath','Full Bath','Half Bath','Bedroom AbvGr','Kitchen AbvGr','TotRms AbvGrd','Fireplaces','Garage Yr Blt','Garage Cars','Mo Sold','Yr Sold']

# Ordinal Features
ordinal_features = ['Lot Shape','Utilities','Land Slope','Overall Qual','Overall Cond','Exter Qual','Exter Cond','Bsmt Qual','Bsmt Cond','Bsmt Exposure','BsmtFin Type 1','BsmtFin Type 2','Heating QC','Electrical','Kitchen Qual','Functional','Fireplace Qu','Garage Finish','Garage Qual','Garage Cond','Paved Drive','Pool QC','Fence']

df_nom = pd.get_dummies(df[nominal_features], columns = nominal_features, drop_first = True)

df_cont = df[continuous_features]
df_dis = df[discrete_features]


X = pd.concat((df_cont, df_dis, df_ord, df_nom), axis = 1)
y = df['SalePrice']

X.shape

# train test split

X_train, X_test, y_train, y_test = train_test_split(X,y, random_state = 42)

# scale model
ss = StandardScaler()
ss.fit(X_train)

X_train = ss.transform(X_train)
X_test = ss.transform(X_test)

## Lasso Model

alphas = np.linspace(0.05,100000, 1000)

lcv = LassoCV(alphas = alphas)
lcv.fit(X_train, y_train)
lcv.score(X_train, y_train), lcv.score(X_test, y_test)

lcv.alpha_

sum(lcv.coef_ != 0)

# from kobe review 

cols_and_coef = pd.DataFrame({ 
    'var': X.columns,
    'coef val': lcv.coef_
}).set_index('var').sort_values('coef val', ascending=False)

cols_and_coef.head()

cols_and_coef[cols_and_coef['coef val'] != 0].index

## Fit to all training data

ss1 = StandardScaler()
ss1.fit(X)

X1 = ss1.transform(X)

lcv2 = LassoCV(alphas = alphas)
lcv2.fit(X1, y)
lcv2.score(X1, y)

lcv2.alpha_

sum(lcv2.coef_ != 0)

X.shape

cols_and_coef = pd.DataFrame({ 
    'var': X.columns,
    'coef val': lcv2.coef_
}).set_index('var').sort_values('coef val', ascending=False)

#export predictions
filepath = '../datasets/interim_files/lasso_fitted_st6.csv'

cols_and_coef.to_csv(filepath)

## do to test

# import cleaned test info
filepath = '../datasets/interim_files/test_clean.csv'

testdata = pd.read_csv(filepath)
testdata.head()

# import ordinal features

filepath = '../datasets/interim_files/testdata_updated_ordinal_features.csv'

testdata_ord = pd.read_csv(filepath)

testdata_nom = pd.get_dummies(testdata[nominal_features], columns = nominal_features, drop_first = True)

testdata_cont = testdata[continuous_features].fillna(0)
testdata_dis = testdata[discrete_features]

Xtestdata = pd.concat((testdata_cont, testdata_dis, testdata_ord, testdata_nom), axis = 1)

df_nom.shape

testdata_nom.shape

df_nom.head()

testdata_nom.head()

###  Because the test data and training data have different dummy columns - I add them together, then separate them, and fill all null values that were created by adding the empty columns with 0 (this way the model will have the same number of columns to operate on for the training and the test data)

overall_nom = pd.concat([df_nom,testdata_nom], join = 'outer', axis = 0)
overall_nom.head()

df_nom_updated = overall_nom[0:2026].fillna(0)

testdata_nom_updated = overall_nom[2026:].fillna(0)

# recalculate lasso cv after updating the columns in the training and test dataframes

X2 = pd.concat((df_cont, df_dis, df_ord, df_nom_updated), axis = 1)
y = df['SalePrice']

ss2 = StandardScaler()
ss2.fit(X2)

X3 = ss2.transform(X2)

lcv3 = LassoCV(alphas = alphas)
lcv3.fit(X3, y)
lcv3.score(X3, y)

lcv3.alpha_

sum(lcv3.coef_ != 0)

var_and_coeff = pd.DataFrame({ 
    'var': X2.columns,
    'coef val': lcv3.coef_
}).set_index('var').sort_values('coef val', ascending=False)

var_and_coeff.head()

var_and_coeff_nonzero = var_and_coeff[var_and_coeff['coef val'] != 0]

var_and_coeff_nonzero

## Apply similar method to test data

Xtestdata2 = pd.concat((testdata_cont, testdata_dis, testdata_ord, testdata_nom_updated), axis = 1)

Xtestdata2 = ss2.transform(Xtestdata2)

y_pred = lcv3.predict(Xtestdata2)

y_pred_df = pd.DataFrame(y_pred)
y_pred_df = y_pred_df.rename(columns={0:'SalePrice'})
y_pred_df.head()

final_preds = pd.concat([testdata['Id'],y_pred_df], axis=1 )
final_preds.info()

#export predictions
filepath = '../datasets/submissions/prediction3.csv'

final_preds.to_csv(filepath, index=False)

## Ridge Model

alphas = np.linspace(0.05,1000, 10)

rcv = RidgeCV(alphas=alphas)
rcv.fit(X_train,y_train)
rcv.score(X_train, y_train), rcv.score(X_test, y_test)

rcv.alpha_

rcv.coef_

# similar results to lasso, but has many more features and , so I am going to ignore

# Notebook Summary:

In this notebook, I use my cleaned training data, then look at all possible features, using all continuous and discrete features, importing the updated ordinal features (the ones with numerical values) from notebook 4, and then creating dummy columns for all nominal features with dropping the first dummy column for each one.

I perform a train/test split on this data, fit a scaler on the train split and then transform both the training and test splits.

I then fit a lasso model on the train split and run the model on both the training and test split.  I get good scores back, so I then perform the scaling and fitting of the lasso model to the entire set of training data.  I then use this updated lasso model on the test data to create **prediction 3**.  Note that while fitting the lasso model to the test data, I realize that some dummy columns don't align with the training and test data, so I combine these, then separate them so that both the training and test data will have the same columns.  All values that are null (because they didn't previously exist, are replaced with 0).

I export the lasso model on the training data for use in other notebooks.


I then fit a ridge model on the train split and run the model on both the training and test split.  However, it has similar performance to the lasso model, and a lot more features, so I don't use it to create any predictions with the test data.
